services/cassandra_communication/requirements.txt
pydantic
cassandra-driver
